requirements.txt
twitch-python
discord.py[voice]
wavelink
python-dotenv
aiosqlite
emoji